==> scripts/02_doublets_detection.qmd <==
# Set seed
```{r}
set.seed(5369553)
```

# Imports
```{r}
suppressPackageStartupMessages({
 library(Seurat)
 library(scDblFinder)
 library(scater) # For the plotUMAP function
})
```

# Load data
```{r}
filtered_data <- readRDS("../data/filtered_data.rds")
```

# Find doublets
```{r}
#Convert your seurat object to a sce object
set.seed(10010101)
sce <- as.SingleCellExperiment(filtered_data)
sce <- scDblFinder(sce,
                   samples = "donor_id",
                   clusters = colLabels(sce))

plotUMAP(sce,
         colour_by = "scDblFinder.score")
ggsave("../results/02/umap_coloured_by_dbl_score.png")

table(sce$scDblFinder.class)

plotColData(sce,
            x = "RNA_snn_res.0.7",
            y = "scDblFinder.score",
            colour_by = ("scDblFinder.class"))
ggsave("../results/02/scatter_dbl_detection_pr_cluster.png")
```

# Subset to not include doublets
```{r}
filtered_data <- sce |>
  as.Seurat() |>
  subset(subset = scDblFinder.class == "singlet")
```

# Update dimensions object
```{r}
dimensions <- readRDS("../data/dimensions.rds")

# Update dimensions overview
dimensions <- dimensions |>
  dplyr::add_row(step = "Remove doublets",
                 n_genes = nrow(filtered_data),
                 n_cells = ncol(filtered_data))
```

# Save objects
```{r}
saveRDS(filtered_data, "../data/filtered_no_doublets_data.rds")
saveRDS(dimensions, "../data/dimensions.rds")
```
