==> scripts/04_cell_type_annotation.qmd <==
---
title: "04_cell_type_annotations"
format: html
editor: source
---

# Set seed
```{r}
set.seed(5369553)
```

### Loading libraries and data

```{r}
obj <- readRDS(file = "../data/augmented_data.rds")


library(Seurat)
library(dplyr)
library(ggplot2)
```

Looking at the basic UMAP on integrated data

```{r}
plot_reduction_01 <- DimPlot(object = obj, reduction = "umap_harmony", 
        group.by = "i.0.1", 
        label = TRUE)

plot_reduction_015 <- DimPlot(object = obj, reduction = "umap_harmony", 
        group.by = "i.0.15", 
        label = TRUE)

plot_reduction_02 <- DimPlot(object = obj, reduction = "umap_harmony", 
        group.by = "i.0.2", 
        label = TRUE)

plot_reduction_025 <- DimPlot(object = obj, reduction = "umap_harmony", 
        group.by = "i.0.25", 
        label = TRUE) +
        labs(x = "UMAP_1", 
             y = "UMAP_2") + ggtitle(NULL)

plot_reduction_03 <- DimPlot(object = obj, reduction = "umap_harmony", 
        group.by = "i.0.3", 
        label = TRUE)

plot_reduction_035 <- DimPlot(object = obj, reduction = "umap_harmony", 
        group.by = "i.0.35", 
        label = TRUE)


ggsave(plot = plot_reduction_025, 
       file = "../results/04/clustering without annotation.jpg",  
       height = 10, 
       width = 20, 
       units = "cm")

```

By eyeballing we find that the clustering of resolution 0.5 is way to high - i.e. we have too many clusters on a PBMC dataset. The cluster that we would suspect to be monocytes (top left) contains 5 different clusters, which will probably be very hard to cluster out into distinct monocyte populations.

Setting `Idents()` to i.0.25

```{r}
Idents(obj) <- "i.0.25"
```


### ID of the top3 differentially expressed genes per cluster. 
```{r}
# Top3 most differentially expressed markers per cluster
diff_markers <- FindAllMarkers(object = obj, 
                               assay = "RNA", 
                               logfc.threshold = 3, 
                               min.pct = 0.25)

top10_diff <- diff_markers |> 
  group_by(cluster) |> 
  slice_max(order_by = avg_log2FC, n = 3)
```


## Dotplot

Dotplot
```{r}
dotplot1 <- DotPlot(object = obj, features = top10_diff$gene, assay = "RNA", scale = TRUE) +
  theme(axis.text.x = element_text(angle = 60)) +
  theme(axis.text.x = element_text(hjust = 1))

ggsave(plot = dotplot1, 
       file = "../results/04/Dotplot top3 genes per annotated cluster.jpg", 
     height = 15, 
    width = 35, 
   units = "cm")
```


## Identification of differentially expressed genes across clusters

*Doing with the FindConservedMarkers function*

Cluster 0

```{r}
FindConservedMarkers(object = obj,
                     ident.1 = 0,
                     assay = "RNA",
                     grouping.var = "disease",
                     min.cells.group = 0.9)
```

So based on this there is no question that cluster 0 is a T-cell cluster as both LTB (lymphotoxin beta) and IL7R (CD127) which are very specific to T-cells are overexpressed.

However it seems as if cluster 1 is a T-cell cluster too. We need to use FindMarkers that differ between these two clusters. As we suspect that cluster 1 is a CD8 cluster we set ident.1 to cluster 1 and ident.2 to cluster 0 and hope to identify overexpression of CD8 and cytotoxic markers.

```{r}
FindMarkers(object = obj, 
            ident.1 = "1", 
            ident.2 = "0")
```

The genes GZMB, GZMH and GNLY as well as LTB are highly associated to cytotoxicity - all of these are significantly upregulated by cluster 1 compared to 0. Even more CD8A is also enhanced in cluster 1 - leading to the conclusion that cluster 1 is CD8+ T cells.

We show this in a violinplot

```{r}
VlnPlot(object = obj, features = c("GZMB", "GZMK", "CD8A", "IL7R"), 
        alpha = 0.1, 
        idents = c("0", "1", "3"))
```

```{r}
FeaturePlot(object = obj, features = c("GZMB", "GZMK", "CD8A", "IL7R"), reduction = "umap_harmony")
```

CD8A clearly maps to cluster 1 with the cytotoxic markers GZM also mapping to cluster 1 and 3. Thus, cluster 1 is CD8+ T cells (CTL)

### Cluster 1 vs. 3

Next thing is to find cluster 3, which is very close to cluster 1. We use Findmarkers on cluster 3 vs. 1. We suspect the cells to be NK cells, as they cluster closely to CTL and NK cells are cytotoxic too.

```{r}
FindMarkers(object = obj, 
            ident.1 = "3", 
            ident.2 = "1", 
            min.pct = 0.1) 
```

CD8A downregulated by cluster 3 vs cluster 1. Also CD3D and CD3G meaning that they are not CD3 gamma-delta T-cells. 
TRAC is also downregulated meaning, that cluster 3 cells are not T cells. FCGR3A (CD16) is overexpressed and so is FCER1G. 
So cluster 3 is assigned to NK cells.

Cluster 5 Identification of conserved markers. We compare to the three other T/NK cell clusters as we want to ascertain the difference between these and cluster 5. The hypothesis is that cells are either NK-T cells or are activated T cells.

```{r}
FindMarkers(object = obj, 
            ident.1 = "5", 
            ident.2 = c("0", "1", "3"), 
            grouping.var = "disease")
```



CD69 is a T-cell activation marker which could lead us to conclude that the cells in cluster 5 are activated T cells. 

Additionally, transscription factors associated to activation are highly expressed in cluster 5. 
Could cells be activated T cells? 

Comparing to the cluster in which we have CD4+ T- cells (cluster 0)

```{r}
FindMarkers(object = obj, 
            ident.1 = "5", 
            ident.2 = "0")
```
Again overexpression of activation markers such as CD69. Additionally we have overexpression of JUNB, FOS and FOSB - transscription factors for activation. 
The cells are activated T cells. 
Woudl have been nice with a CITE-Seq data to ascertain the phenotype of cells. 




```{r}
lymfoid_violin_plot <- VlnPlot(object = obj, 
        features = c("TRAC", "CD8A", "FCGR3A", "NCAM1", "GZMB", "CD69"), 
        idents = c("1", "0", "3", "5"), 
        alpha = 0.1)

ggsave(plot = lymfoid_violin_plot, 
       file = "../results/04/Violin plot of lymphoid cell features for clustering.jpg", 
       width = 20, 
       height = 10, 
       units = "cm")

```
CD8A expressed almost solely by cluster 1 - CD8+ T-cells
SELL expressed almost solely by cluster 0 - SELL is expressed by T cells - mainly naive. And as cluster 0 is not CD8 but are T-cells they must be CD4+ T-cells. 
CD69 - an activation marker - expressed both by cluster 3 and 5. So these cells are activated. 
FCGR3A and GZMB mainly expressed by cluster 3 - FCGR3A is CD16, which is not expressed by T cells, but by NK- cells so cluster 3 is NK cells. 
FOS - an activation marker is mainly expressed by cluster 5. So cluster 5 is activated T cells. 

Cluster 7
```{r}
FindConservedMarkers(object = obj, 
                     ident.1 = "7", 
                     grouping.var = "disease")
```

Plotting some of the top DE genes in a violinplot

```{r}
VlnPlot(object = obj, features = c("PPBP", "GP1BB", "PF4"))
```

Both PPBP, GP1BB and PF4 are highly expressed by platelets, so cluster 7 is platelets.

Cluster 8

```{r}
FindConservedMarkers(object = obj, 
                     ident.1 = "8", 
                     grouping.var = "disease")
```

SERPINF1 is expressed by pDC. SCT is only expressed by pDC. 
Cluster 8 - DC. 

Cluster 2 "overshadows" what is going on in cluster 8, so let's try do FindMarkers in cluster 8 vs. cluster 6.

Plotting DC markers on DC clusters and monocyte clusters. 
```{r}
VlnPlot(object = obj, 
        features = c("ITM2C", "ALOX5AP", "SERPINF1", "TCF4"), 
        idents = c("8", "2"))
```

So cluster 8 is DC (probably both mDC and pDC).

Cluster 4

```{r}
FindConservedMarkers(object = obj, ident.1 = "4", grouping.var = "disease")
```

CD79A and CD83 are expressed by B cells. So is MS4A1

```{r}
VlnPlot(object = obj, features = c("CD83", "CD79A", "MS4A1"))
```

CD83 is actually expressed by several clusters and cells (monocytes and DC), byt only CD79A and MS4A1 are expressed by cluster 4, so cluster 4 has to be B-cells.

Cluster 2

```{r}
FindConservedMarkers(object = obj, 
                     ident.1 = "2", 
                     grouping.var = "disease")
```

LYZ is expressed by both monocytes and mDC. So is FCN1, IFI30 and FCER1G. Could be that mDC are within the cluster of monocytes.

We have to find the difference between cluster 2 and 6.

```{r}
FindMarkers(object = obj, 
            ident.1 = "2", 
            ident.2 = "6")
```

FCGR3A is actually CD16, which is an immunoglobulin receptor that mediates opsonization - a non-classical monocyte feature. Additional FCGR3A is overexpressed by non-classical monocytes compared to classical monocytes according to the human protein atlas. IFITM2 is also overexpressed by non-classical monocytes as is LST1.

We can just check CD14 (classical monocyte marker) and FCGR3A (CD16) expression in a FeaturePlot. FCER1G is a classical monocyte marker.

```{r}
FeaturePlot(object = obj, features = c("CD14", "FCER1G", "FCGR3A"), reduction = "umap_harmony")
```

We find that cells in the cluster 3 also expresses CD16 - the NK cells.

So cluster 2 is classical monocytes. Cluster 6 is non-classical monocytes.



```{r}
## Checking number of cells per cluster
table(obj@meta.data$i.0.25)
```

Setting the cell identities and setting the cell identities to the column "dtu_cell_type"

```{r}
Idents(obj) <- "i.0.25"

obj <- RenameIdents(object = obj, 
                    "0" = "CD4+ T-cells",
                    "1" = "CD8+ T-cells",
                    "2" = "Classical monocytes",
                    "3" = "NK cells",
                    "4" = "B cells",
                    "5" = "Activated T-cells",
                    "6" = "Non-classical monocytes",
                    "7" = "Platelets",
                    "8" = "DC")


## Adding the cell IDs to the metadata as a new column - is named cell_type_manual
obj[["dtu_cell_type"]] <- Idents(object = obj)
```

Checking in UMAP

```{r}
umap_w_annotations <- DimPlot(object = obj,
        reduction = "umap_harmony",
        label = TRUE,
        group.by = "dtu_cell_type") + 
  ggtitle(NULL) + 
  labs(x = "UMAP_1", 
       y = "UMAP_2") + NoLegend()

ggsave(plot = umap_w_annotations, 
       filename = "../results/04/UMAP cell 04s.jpg", height = 10, width = 20, units = "cm")
```


### Heatmap
Not used
```{r}

#heatmap_top_3_genes <- Seurat::DoHeatmap(object = obj, 
 #                                        features = top10_diff$gene,
  #                                       cells = 1:500,
   #                                      angle = 90, 
    #              size = 4, 
     #             label = FALSE, 
      #            slot = "scale.data")

```


Featureplots with the best marker for each cluster
```{r}
featureplots_all_clusters <- FeaturePlot(object = obj, 
            features = c("SELL", "CD8B", "KLRF1", "CD14", "CD79A", "CD69", "FCGR3A", "PPBP", "SCT"), 
            reduction = "umap_harmony")

ggsave(plot = featureplots_all_clusters, 
       file = "../results/04/Featureplot for all clusters.jpg", 
       width = 25, 
       height = 15, 
       units = "cm", 
       dpi = 600)
```




Saving file with annotated clusters

```{r}
saveRDS(object = obj, 
        file = "../data/annotated_data.rds")
```
