==> scripts/06_cell_type_abundance.qmd <==
```{r}
suppressPackageStartupMessages({
library(DCATS)
library(Seurat)
library(SeuratData)
library(dplyr)
library(enrichplot)
library(ggplot2)
library(ggrepel)
library(tibble)
library(tidyverse)
})
```


```{r}
data <- readRDS("../data/annotated_data.rds")
```

```{r}
data[["case_control"]] <- data@meta.data$CHIP
```

```{r}
# Convert case_control to a character vector to allow modifications
data$case_control <- as.character(data$case_control)

# Modify the case_control levels
data$case_control[data$case_control == "control"] <- "control"
data$case_control[data$case_control == "chip"] <- "case"
# View(data)
# Optionally, convert CHIP back to a factor
data$case_control <- factor(data$case_control, levels = c("control", "case"))

# Add id to condition
data$case_control_id <- paste0(data$case_control, data$donor_id)
```

```{r}
count_mat = table(data$case_control_id, data$dtu_cell_type)
```

```{r}
# Calculate the similarity matrix using the KNN graph
knn_mat <- knn_simMat(
  data@graphs$RNA_snn,  # Use Seurat's SNN graph
  data$dtu_cell_type    # Specify the cell types
)
```

```{r}
# Create the design matrix
design_mat <- data.frame(
  condition = ifelse(grepl("control", rownames(count_mat)), "ahealthy", "disease")
)

rownames(design_mat) <- rownames(count_mat)
design_mat
```

```{r}
# Perform DCATS analysis
dcats_results <- dcats_GLM(
  count_mat = count_mat,
  design_mat = design_mat,
  similarity_mat = knn_mat
)
dcats_results
```

```{r}
# Extract and inspect results
dcats_results$fdr
```

```{r}
# rowSums(count_mat)  # Summarize counts for each cell type
# colSums(count_mat)  # Summarize counts for each condition
# table(design_mat$condition)

```

```{r}
# # Visualize the FDR results (Optional)
fdr_df <- as.data.frame(
                        dcats_results$fdr,
                        ) |> 
  tibble::as_tibble(rownames = "Cell Type") |> 
  rename(fdr = condition)

coef_df <- as.data.frame(dcats_results$ceoffs)|> 
  tibble::as_tibble(rownames = "Cell Type") |> 
  rename(coeffs = condition)

df <- fdr_df |> left_join(coef_df) |> 
  mutate(sig = case_when(fdr <= 0.05 ~ "Significant",
                         TRUE ~ "Not significant"))

p1 <- ggplot(df,
             aes(x = reorder(`Cell Type`, coeffs),
                 y = coeffs,
                 fill = sig)) +
  geom_bar(stat = "identity") +
    labs(title = "DCATS Results",
       x = "Cell Type",
       y = "Estimated coefficient",
       fill = "")+
  coord_flip()+
  scale_fill_viridis_d() +
  theme_bw()
  
p1

output_dir <- "../results/06"

ggsave(
  filename = file.path(output_dir, "dcats.png"),
  plot = p1,
  width = 8, 
  height = 6,
  dpi = 300,
  scale = 1
)
```
